+++ src/component/common/Form/index.jsx
@@ -1,0 +1,4 @@
+import CommentForm from "./CommentForm/index";
+import ProfileForm from "./ProfileForm/index";
+
+export { CommentForm, ProfileForm };
